+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.3] - 2022-05-17
+## [unreleased]
 
 ### Added
 
 - handle `response_mode` in hybrid requests
+
+### Changed
+
+- prefer `invalid_grant` to `invalid_code` and `invalid_refresh_token`
+- error messages have been improved
+
+## [2.1.3] - 2022-05-17
+
+### Added
+
+- handle `response_mode` in hybrid requests errors
 
 ### Fixed